AI_Agen/AI_Agen/ai_job_finder/requirements.txt
fastapi
uvicorn
requests
streamlit
google-genai
tavily-python
python-dotenv
pandas
